30/06

+++ Backend/opositest/src/main/java/edu/xtd/opositest/modelo/Test.java
@@ -1,5 +1,5 @@
 package edu.xtd.opositest.modelo;
 
-public record Test(long id, String titulo, int numeroDePreguntas) {
+public record Test(int id, String titulo, int numeroDePreguntas) {
 
 }
